feat(bonuses): Basic startingBonuses
Custom bonuses can be added to the game by adding them to the initial
config. A rudimentary timing system enables duration but does not pause
the game.

A handful of bonuses are included in the presets file along with their
CSS.

The Ball now autocorrects its angle when that's too flat.